Add index for better perf on title or author queries.
This index allows us to compare trigrams of the concated text, which gives us %foo% queries actually being indexed (whereas before we would do a full table scan).

This takes us from now where we have ~6k puzzles where a query will take >100ms to <10ms with the index even when it fetches a lot of data (think %sunday%).

This does require us to manually construct the and clauses as ALL is not optimized in the same way and separate and clauses (ALL forces us back to a full table scan).